Extract to object classes
