+++ libcool_gl/include/cool_gl/Window.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace cool_gl {
+
+struct Window {
+
+
+};
+
+} // namespace cool_gl
