+++ libcool_gl/src/Drawable.cpp
@@ -1,0 +1,7 @@
+#include <cool_gl/Drawable.h>
+
+namespace cool_gl {
+
+Drawable::~Drawable(){};
+
+}
