+++ libcool_gl/src/Window.cpp
@@ -1,0 +1,2 @@
+#include <cool_gl/Window.h>
+
